+++ src/fqlite/descriptor/TableDescriptor.java
@@ -51,13 +51,13 @@ public class TableDescriptor extends AbstractDescriptor implements Comparable<Ta
 	
 			/* interpret all byte values as a list of varints */
 			/* each varint represents a columntype */
-			int[] values = Auxiliary.readVarInt(bcol);
+			long[] values = Auxiliary.readVarInt(bcol);
 	
 			/*
 			 * normally, the first byte of the match holds total length of header bytes
 			 * including this byte
 			 */
-			int headerlength = values[0];
+			int headerlength = (int) values[0];
 			info(Integer.toString(headerlength));
 			
 			boolean valid = true;
